Implement batch computation and logging for error bound

+++ MIPVerify/src/batch_processing_helpers.jl
@@ -1,4 +1,4 @@
-export batch_find_untargeted_attack
+export batch_find_untargeted_attack, batch_find_error_bound
 
 @enum SolveRerunOption never=1 always=2 resolve_ambiguous_cases=3 refine_insecure_cases=4 retarget_infeasible_cases=5
 
@@ -11,7 +11,7 @@ end
 
 # BatchRunParameters are used for result folder names
 function Base.show(io::IO, t::BatchRunParameters)
-    print(io, 
+    print(io,
         "$(t.nn.UUID)__$(t.pp)__$(t.norm_order)__$(t.tolerance)"
     )
 end
@@ -48,7 +48,7 @@ end
 
 function generate_csv_summary_line(sample_number::Integer, results_file_relative_path::String, r::Dict)
     [
-        sample_number, 
+        sample_number,
         results_file_relative_path,
         r[:PredictedIndex],
         r[:TargetIndexes],
@@ -65,7 +65,7 @@ end
 function generate_csv_summary_line_optimal(sample_number::Integer, d::Dict)
     assert(d[:PredictedIndex] in d[:TargetIndexes])
     [
-        sample_number, 
+        sample_number,
         "",
         d[:PredictedIndex],
         d[:TargetIndexes],
@@ -93,6 +93,31 @@ function create_summary_file_if_not_present(summary_file_path::String)
             "ObjectiveBound",
             "TighteningApproach",
             "TotalTime"
+        ]
+
+        open(summary_file_path, "w") do file
+            writecsv(file, [summary_header_line])
+        end
+    end
+end
+
+function create_summary_file_if_not_present_for_error(summary_file_path::String)
+    if !isfile(summary_file_path)
+        summary_header_line = [
+            "SampleNumber",
+            "OffsetMin",
+            "OffsetMax",
+            "OffsetMinSolved",
+            "OffsetMinStatus",
+            "OffsetMaxSolved",
+            "OffsetMaxStatus",
+            "AngleMin",
+            "AngleMax",
+            "AngleMinSolved",
+            "AngleMinStatus",
+            "AngleMaxSolved",
+            "AngleMaxStatus",
+            "SolveTime",
         ]
 
         open(summary_file_path, "w") do file
@@ -132,9 +157,30 @@ function initialize_batch_solve(
 
     summary_file_path = joinpath(main_path, summary_file_name)
     summary_file_path|> create_summary_file_if_not_present
-    
+
     dt = CSV.read(summary_file_path)
     return (results_dir, main_path, summary_file_path, dt)
+end
+
+function initialize_batch_solve_for_error(
+    save_path::String,
+    offset_grid_num::Integer,
+    angle_grid_num::Integer,
+    grid_size::Real,
+    )::Tuple{String,String,DataFrames.DataFrame}
+
+    summary_file_name = "summary.csv"
+    batch_run_parameters = "offset_$(offset_grid_num)_angle_$(angle_grid_num)_grid_size_$(grid_size)"
+
+    main_path = joinpath(save_path, batch_run_parameters |> string)
+
+    main_path |> mkpath_if_not_present
+
+    summary_file_path = joinpath(main_path, summary_file_name)
+    summary_file_path|> create_summary_file_if_not_present_for_error
+
+    dt = CSV.read(summary_file_path)
+    return (main_path, summary_file_path, dt)
 end
 
 function save_to_disk(
@@ -162,6 +208,33 @@ function save_to_disk(
     end
 end
 
+function save_to_csv_for_error(
+    sample_number::Integer,
+    summary_file_path::String,
+    d::Dict
+    )
+    summary_line = [
+        sample_number,
+        d[:OffsetMin],
+        d[:OffsetMax],
+        d[:OffsetMinSolved],
+        d[:OffsetMinStatus],
+        d[:OffsetMaxSolved],
+        d[:OffsetMaxStatus],
+        d[:AngleMin],
+        d[:AngleMax],
+        d[:AngleMinSolved],
+        d[:AngleMinStatus],
+        d[:AngleMaxSolved],
+        d[:AngleMaxStatus],
+        d[:SolveTime],
+    ] .|> string
+    open(summary_file_path, "a") do file
+        writecsv(file, [summary_line])
+    end
+end
+
+
 """
 $(SIGNATURES)
 
@@ -170,16 +243,16 @@ looking up information on the most recent completed solve recorded in `summary_d
 matching `sample_number`.
 
 `summary_dt` is expected to be a `DataFrame` with columns `:SampleNumber`, `:SolveStatus`,
-and `:ObjectiveValue`. 
+and `:ObjectiveValue`.
 
 Behavior for different choices of `solve_rerun_option`:
 + `never`: `true` if and only if there is no previous completed solve.
 + `always`: `true` always.
-+ `resolve_ambiguous_cases`: `true` if there is no previous completed solve, or if the 
++ `resolve_ambiguous_cases`: `true` if there is no previous completed solve, or if the
     most recent completed solve a) did not find a counter-example BUT b) the optimization
     was not demosntrated to be infeasible.
 + `refine_insecure_cases`: `true` if there is no previous completed solve, or if the most
-    recent complete solve a) did find a counter-example BUT b) we did not reach a 
+    recent complete solve a) did find a counter-example BUT b) we did not reach a
     provably optimal solution.
 """
 function run_on_sample_for_untargeted_attack(sample_number::Integer, summary_dt::DataFrame, solve_rerun_option::MIPVerify.SolveRerunOption)::Bool
@@ -210,23 +283,23 @@ end
 $(SIGNATURES)
 
 Runs [`find_adversarial_example`](@ref) for the specified neural network `nn` and `dataset`
-for samples identified by the `target_indices`, with the target labels for each sample set 
+for samples identified by the `target_indices`, with the target labels for each sample set
 to the complement of the true label.
-    
-It creates a named directory in `save_path`, with the name summarizing 
-  1) the name of the network in `nn`, 
-  2) the perturbation family `pp`, 
+
+It creates a named directory in `save_path`, with the name summarizing
+  1) the name of the network in `nn`,
+  2) the perturbation family `pp`,
   3) the `norm_order`
   4) the `tolerance`.
 
-Within this directory, a summary of all the results is stored in `summary.csv`, and 
+Within this directory, a summary of all the results is stored in `summary.csv`, and
 results from individual runs are stored in the subfolder `run_results`.
 
 This functioned is designed so that it can be interrupted and restarted cleanly; it relies
 on the `summary.csv` file to determine what the results of previous runs are (so modifying
 this file manually can lead to unexpected behavior.)
 
-If the summary file already contains a result for a given target index, the 
+If the summary file already contains a result for a given target index, the
 `solve_rerun_option` determines whether we rerun [`find_adversarial_example`](@ref) for this
 particular index.
 
@@ -236,10 +309,10 @@ particular index.
 + `save_path`: Directory where results will be saved. Defaults to current directory.
 + `pp, norm_order, tolerance, rebuild, tightening_algorithm, tightening_solver, cache_model,
   solve_if_predicted_in_targeted` are passed
-  through to [`find_adversarial_example`](@ref) and have the same default values; 
+  through to [`find_adversarial_example`](@ref) and have the same default values;
   see documentation for that function for more details.
-+ `solve_rerun_option::MIPVerify.SolveRerunOption`: Options are 
-  `never`, `always`, `resolve_ambiguous_cases`, and `refine_insecure_cases`. 
++ `solve_rerun_option::MIPVerify.SolveRerunOption`: Options are
+  `never`, `always`, `resolve_ambiguous_cases`, and `refine_insecure_cases`.
   See [`run_on_sample_for_untargeted_attack`](@ref) for more details.
 """
 function batch_find_untargeted_attack(
@@ -259,7 +332,7 @@ function batch_find_untargeted_attack(
     solve_if_predicted_in_targeted = true,
     adversarial_example_objective::AdversarialExampleObjective = closest
     )::Void
-    
+
     verify_target_indices(target_indices, dataset)
     (results_dir, main_path, summary_file_path, dt) = initialize_batch_solve(save_path, nn,  pp, norm_order, tolerance)
 
@@ -359,7 +432,7 @@ function batch_find_targeted_attack(
                 end
 
                 info(MIPVerify.LOGGER, "Working on index $(sample_number), with true_label $(true_one_indexed_label) and target_label $(target_label)")
-            
+
                 d = find_adversarial_example(nn, input, target_label, main_solver, invert_target_selection = false, pp=pp, norm_order=norm_order, tolerance=tolerance, rebuild=rebuild, tightening_algorithm = tightening_algorithm, tightening_solver = tightening_solver, cache_model=cache_model, solve_if_predicted_in_targeted=solve_if_predicted_in_targeted)
 
                 save_to_disk(sample_number, main_path, results_dir, summary_file_path, d, solve_if_predicted_in_targeted)
@@ -384,6 +457,92 @@ function batch_build_model(
         info(MIPVerify.LOGGER, "Working on index $(sample_number)")
         input = MIPVerify.get_image(dataset.images, sample_number)
         build_reusable_model_uncached(nn, input, pp, tightening_solver, tightening_algorithm)
+    end
+    return nothing
+end
+
+
+function batch_find_error_bound(
+    nn::NeuralNet,
+    dataset::MIPVerify.RangeDataset,
+    main_solver::MathProgBase.SolverInterface.AbstractMathProgSolver;
+    save_path::String = ".",
+    solve_rerun_option::MIPVerify.SolveRerunOption = MIPVerify.never,
+    pp::MIPVerify.PerturbationFamily = MIPVerify.CustomPerturbationFamily(),
+    tightening_algorithm::MIPVerify.TighteningAlgorithm = DEFAULT_TIGHTENING_ALGORITHM,
+    tightening_solver::MathProgBase.SolverInterface.AbstractMathProgSolver = MIPVerify.get_default_tightening_solver(main_solver),
+    )::Void
+
+    (main_path, summary_file_path, dt) = initialize_batch_solve_for_error(
+        save_path,
+        dataset.offset_grid_num,
+        dataset.angle_grid_num,
+        dataset.grid_size
+        )
+    num_entries = MIPVerify.num_samples(dataset)
+
+    non_optimal_entries = []
+    for sample_number in 1:num_entries
+        if run_on_sample_for_untargeted_attack(sample_number, dt, solve_rerun_option)
+            info(MIPVerify.LOGGER, "Working on index $(sample_number)")
+
+            input_lower_bound = dataset.image_lower_bounds[sample_number:sample_number,:,:,:]
+            input_upper_bound = dataset.image_upper_bounds[sample_number:sample_number,:,:,:]
+            offset_low = dataset.offset_lower_bounds[sample_number]
+            offset_high = dataset.offset_upper_bounds[sample_number]
+            angle_low = dataset.angle_lower_bounds[sample_number]
+            angle_high = dataset.angle_upper_bounds[sample_number]
+            summary_item = Dict()
+            summary_item[:OffsetMin] = offset_low
+            summary_item[:OffsetMax] = offset_high
+            summary_item[:AngleMin] = angle_low
+            summary_item[:AngleMax] = angle_high
+
+            result_dict = MIPVerify.find_range_for_outputs(
+                nn,
+                input_lower_bound,
+                input_upper_bound,
+                main_solver,
+                pp = pp,
+                tightening_algorithm = tightening_algorithm,
+                tightening_solver = tightening_solver,
+            )
+            summary_item[:OffsetMinSolved] = result_dict[:OffsetMin][:ObjectiveValue]
+            summary_item[:OffsetMaxSolved] = result_dict[:OffsetMax][:ObjectiveValue]
+            summary_item[:AngleMinSolved] = result_dict[:AngleMin][:ObjectiveValue]
+            summary_item[:AngleMaxSolved] = result_dict[:AngleMax][:ObjectiveValue]
+            summary_item[:OffsetMinStatus] = result_dict[:OffsetMin][:SolveStatus]
+            summary_item[:OffsetMaxStatus] = result_dict[:OffsetMax][:SolveStatus]
+            summary_item[:AngleMinStatus] = result_dict[:AngleMin][:SolveStatus]
+            summary_item[:AngleMaxStatus] = result_dict[:AngleMax][:SolveStatus]
+            summary_item[:SolveTime] = result_dict[:TotalTime]
+            save_to_csv_for_error(sample_number, summary_file_path, summary_item)
+            for setting in [:OffsetMin, :OffsetMax, :AngleMin, :AngleMax]
+                if result_dict[setting][:SolveStatus] != :Optimal
+                    append!(non_optimal_entries, sample_number)
+                    break
+                end
+            end
+        end
+    end
+
+    # Get error matrix
+    summary_dt = CSV.read(summary_file_path)
+    offset_errors = max(summary_dt[:OffsetMaxSolved] - summary_dt[:OffsetMin], summary_dt[:OffsetMax] - summary_dt[:OffsetMinSolved])
+    angle_errors = max(summary_dt[:AngleMaxSolved] - summary_dt[:AngleMin], summary_dt[:AngleMax] - summary_dt[:AngleMinSolved])
+    error_result = Dict()
+    error_result["offset_errors"] = offset_errors
+    error_result["angle_errors"] = angle_errors
+    error_result["offset_grid_num"] = dataset.offset_grid_num
+    error_result["angle_grid_num"] = dataset.angle_grid_num
+    matwrite(joinpath(main_path, "error_bound_result.mat"), error_result)
+
+    # Report if any entry is not solved to optimal
+    if !isempty(non_optimal_entries)
+        println("There are samples not solved to optimal:\n")
+        println(non_optimal_entries)
+    else
+        println("All samples are solved to optimal.")
     end
     return nothing
 end
